👔enhancement: 장바구니 페이지 기능개선 작업(Issue #92 해결)

+++ src/pages/Carts.jsx
@@ -3,16 +3,17 @@ import CartsBox from "@components/carts/CartsBox";
 import CartsDelete from "@components/carts/CartsDelete";
 import CartsPayment from "@components/carts/CartsPayment";
 import useAxiosInstance from "@hooks/useAxiosInstance";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { Helmet } from "react-helmet-async";
 
 export default function Carts() {
   const axios = useAxiosInstance();
   const [loading, setLoading] = useState(true); // 로딩 상태
   const [error, setError] = useState(null); // 에러 상태
-  const [carts, setCarts] = useState([]); // 장바구니 정보
+  const [carts, setCarts] = useState({ item:[] }); // 장바구니 정보
   const [checkedItems, setCheckedItems] = useState([]); // 선택된 항목의 배열
   const [allChecked, setAllChecked] = useState(true); // 전체 선택 상태
+  const [product, setProduct] = useState(null); // 상품 초기값 null
 
   // 장바구니 상품 수량 수정 (/carts/{_id})
   const patchQuantityPlusCart = async _id => {
@@ -61,7 +62,6 @@ export default function Carts() {
       }));
     }
   };
-
   // 체크 상태 변경
   const handleCheckboxChange = id => {
     const newCheckedItems = checkedItems.includes(id)
@@ -73,7 +73,6 @@ export default function Carts() {
     // 전체 선택 상태 업데이트
     setAllChecked(newCheckedItems.length === carts.length);
   };
-
   // 전체 선택 체크박스 변경
   const handleAllCheckboxChange = () => {
     if (allChecked) {
@@ -84,7 +83,11 @@ export default function Carts() {
     setAllChecked(!allChecked); // 전체 선택 상태 반전
   };
   // 장바구니 상품 한건 삭제 (/carts/{_id})
-  const DeleteCarts = async _id => {
+  const deleteCarts = async _id => {
+    const deleteCartsConfirm = window.confirm(
+      "해당 상품을 장바구니에서 제거 하시겠습니까?",
+    );
+    if (!deleteCartsConfirm) return;
     setLoading(true); // 삭제 요청 시작 시 로딩 상태 설정
     try {
       await axios.delete(`/carts/${_id}`);
@@ -110,52 +113,69 @@ export default function Carts() {
       setLoading(false); // 로딩 종료
     }
   };
+  
+  // 상품 상세 조회 (/products/{_id})
+  const fetchProduct = async () => {
+    try {
+      const response = await axios.get(`/products/`);
+      setProduct(response?.data);
+    } catch (err) {
+      setError(err);
+    }
+  };
+    useEffect(() => {
+      fetchProduct();
+      fetchCarts(); // 장바구니 정보 가져오기
+    }, []);
+    useEffect(() => {
+      if (carts?.item) {
+        setCheckedItems(carts.item.map(cartlist => cartlist._id)); // 모든 체크박스를 선택됨으로 설정
+      }
+    }, [product]);
+  
   return (
     <>
-      <Helmet>
-        <title>장바구니 - ILAND</title>
+    <Helmet>
+    <title>장바구니 - ILAND</title>
 
-        <meta property="og:title" content="장바구니 - ILAND" />
-        <meta
-          property="og:description"
-          content="ILAND에서 내 취향을 모아보세요."
-        />
-      </Helmet>
-      <div className="container">
-        <CartsDelete
-          setCarts={setCarts}
-          setError={setError}
-          setLoading={setLoading}
-          checkedItems={checkedItems}
-          handleAllCheckboxChange={handleAllCheckboxChange}
-          allChecked={allChecked}
-        />
-        {carts.item?.length === 0 ? (
-          <CartEmpty />
-        ) : (
-          <>
-            <CartsBox
-              error={error}
-              loading={loading}
-              setError={setError}
-              setLoading={setLoading}
-              setCheckedItems={setCheckedItems}
-              fetchCarts={fetchCarts}
-              carts={carts}
-              handleCheckboxChange={handleCheckboxChange}
-              checkedItems={checkedItems}
-              patchQuantityPlusCart={patchQuantityPlusCart}
-              patchQuantityMinusCart={patchQuantityMinusCart}
-              DeleteCarts={DeleteCarts}
-            />
-            <CartsPayment
-              checkedItems={checkedItems}
-              setCarts={setCarts}
-              carts={carts}
-            />
-          </>
-        )}
-      </div>
+    <meta property="og:title" content="장바구니 - ILAND" />
+    <meta
+      property="og:description"
+      content="ILAND에서 내 취향을 모아보세요."
+    />
+  </Helmet>
+    <div className="container">
+      <CartsDelete
+        setCarts={setCarts}
+        setError={setError}
+        setLoading={setLoading}
+        checkedItems={checkedItems}
+        handleAllCheckboxChange={handleAllCheckboxChange}
+        allChecked={allChecked}
+      />
+      {carts.item?.length === 0 ? (
+        <CartEmpty />
+      ) : (
+        <>
+          <CartsBox
+            error={error}
+            loading={loading}
+            carts={carts?.item}
+            product={product}
+            checkedItems={checkedItems}
+            handleCheckboxChange={handleCheckboxChange}
+            patchQuantityPlusCart={patchQuantityPlusCart}
+            patchQuantityMinusCart={patchQuantityMinusCart}
+            deleteCarts={deleteCarts}
+          />
+          <CartsPayment
+            checkedItems={checkedItems}
+            setCarts={setCarts}
+            carts={carts.item}
+          />
+        </>
+      )}
+    </div>
     </>
   );
 }
